+++ src/App.tsx
@@ -6,17 +6,19 @@ import Nav from './components/Nav';
 // Containers
 import Home from './containers/Home/';
 import About from './containers/About/';
+import Contact from './containers/Contact/';
 import CreateDoc from './containers/CreateDoc/';
 
 const App: React.FC = () => (
-    <Router>
-        <Nav />
-        <main className="page_main">
-            <Route exact path="/" component={Home} />
-            <Route path="/about" component={About} />
-            <Route path="/crear-documento" component={CreateDoc} />
-        </main>
-    </Router>
+  <Router>
+    <Nav />
+    <main className="page_main">
+      <Route exact path="/" component={Home} />
+      <Route path="/about" component={About} />
+      <Route path="/contact" component={Contact} />
+      <Route path="/crear-documento" component={CreateDoc} />
+    </main>
+  </Router>
 );
 
 export default App;
